trim down implementation classes and make them inner
